backend/requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiogram-unittest==0.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
coverage==7.6.9
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dotenv==1.0.1
python-multipart==0.0.19
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
WTForms==3.1.2
yarl==1.18.3
